+++ app/src/main/res/menu/layout_navigation_drawer.xml
@@ -14,6 +14,8 @@
                 android:title="Share" />
             <item android:id="@+id/nav_send" android:icon="@android:drawable/ic_menu_send"
                 android:title="Send" />
+            <item android:id="@+id/nav_logout" android:icon="@android:drawable/ic_menu_close_clear_cancel"
+                android:title="@string/menu_logout" />
         </menu>
     </item>
