+++ app/src/main/res/values/strings.xml
@@ -4,6 +4,7 @@
 
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
+    <string name="menu_logout">Cerrar sesión</string>
 
     <string name="action_settings">Settings</string>
     <string name="title_activity_navigation">NavigationActivity</string>
